Updated the error jsps, added an erroNav jsp when an error occurs, updated the CSS for the jsp. Added logic to check for a current session and routed to an error page if a logged out user tries to access the webapp witout being logged in. Added comments and JavaDoc.

+++ src/main/java/edu/matc/controller/DeleteMeal.java
@@ -2,15 +2,17 @@ package edu.matc.controller;
 
 import edu.matc.entity.UserFood;
 import edu.matc.persistence.GenericDao;
+
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+
 import java.io.IOException;
 
 /**
@@ -34,13 +36,23 @@ public class DeleteMeal extends HttpServlet {
      */
     @Override
     public void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        // Get the which _method was submitted with the form
-        String method = request.getParameter("_method");
+        // Get the session and sub string from the session
+        HttpSession session = request.getSession(false);
+        String sub = session.getAttribute("sub").toString();
 
-        // If the method was a DELETE request, then call the doDelete method and pass in the request and response
-        if (method.equals("DELETE")) {
-            doDelete(request, response);
+        // If the session is not null and the sub string is not null, continue on. Else route to error page
+        if (session != null && sub != null) {
+            // Get the which _method was submitted with the form
+            String method = request.getParameter("_method");
+
+            // If the method was a DELETE request, then call the doDelete method and pass in the request and response
+            if (method.equals("DELETE")) {
+                doDelete(request, response);
+            }
+        } else {
+            // TODO: Route to error page
         }
+
     }
 
     /**
@@ -53,41 +65,50 @@ public class DeleteMeal extends HttpServlet {
      */
     @Override
     public void doDelete(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        // Instantiate the GenericDao with the UserFood entity
-        GenericDao<UserFood> genericDao = new GenericDao<>(UserFood.class);
+        // Get the session and sub string from the session
+        HttpSession session = request.getSession(false);
+        String sub = session.getAttribute("sub").toString();
 
-        try {
-            // Get the meal id from the parameter food_to_delete
-            String mealID = request.getParameter("food_to_delete");
+        // If the session is not null and the sub string is not null, continue on. Else route to error page
+        if (session != null && sub != null) {
+            // Instantiate the GenericDao with the UserFood entity
+            GenericDao<UserFood> genericDao = new GenericDao<>(UserFood.class);
 
-            // Log id for debugging to make sure correct id is captured
-            logger.debug("mealToDelete: " + mealID);
+            try {
+                // Get the meal id from the parameter food_to_delete
+                String mealID = request.getParameter("food_to_delete");
 
-            // If mealID is not null
-            if (mealID != null) {
-                // Parse mealID to an integer
-                int mealToDelete = Integer.parseInt(mealID);
+                // Log id for debugging to make sure correct id is captured
+                logger.debug("mealToDelete: " + mealID);
 
-                // Use the genericDao to get the meal by passing in the id
-                UserFood foodToDelete = genericDao.getById(mealToDelete);
+                // If mealID is not null
+                if (mealID != null) {
+                    // Parse mealID to an integer
+                    int mealToDelete = Integer.parseInt(mealID);
 
-                // Log meal to make sure the correct meal is being deleted for debugging
-                logger.debug("foodToDelete: " + foodToDelete);
+                    // Use the genericDao to get the meal by passing in the id
+                    UserFood foodToDelete = genericDao.getById(mealToDelete);
 
-                // If foodToDelete exists and does not return null
-                if (foodToDelete != null) {
-                    // Call the deleteEntity method on the genericDao and pass in the meal id
-                    genericDao.deleteEntity(foodToDelete);
+                    // Log meal to make sure the correct meal is being deleted for debugging
+                    logger.debug("foodToDelete: " + foodToDelete);
 
-                    // Log the success message for debugging
-                    logger.debug("Successfully deleted Meal: " + foodToDelete);
+                    // If foodToDelete exists and does not return null
+                    if (foodToDelete != null) {
+                        // Call the deleteEntity method on the genericDao and pass in the meal id
+                        genericDao.deleteEntity(foodToDelete);
+
+                        // Log the success message for debugging
+                        logger.debug("Successfully deleted Meal: " + foodToDelete);
+                    }
                 }
-            }
 
-            // Redirect user to the /meal-display route to view changes
-            response.sendRedirect(request.getContextPath() + "/meal-display");
-        } catch (Exception e) {
-            logger.error("Error while deleting meal", e);
+                // Redirect user to the /meal-display route to view changes
+                response.sendRedirect(request.getContextPath() + "/meal-display");
+            } catch (Exception e) {
+                logger.error("Error while deleting meal", e);
+            }
+        } else {
+            // TODO: Route to error page
         }
     }
 }
